+++ CHANGES.md
@@ -1,5 +1,9 @@
 #### [unreleased]
 
+#### 8.8.1 / 2019-06-11
+* set `homepage` to `PluginURI` or `ThemeURI`, fixes [#791](https://github.com/afragen/github-updater/issues/791)
+* fixed Bitbucket release asset updates for proper containing folder structure, thanks @benoitchantre for the bug report
+
 #### 8.8.0 / 2019-05-15
 * switched from `pre_set_site_transient_update_{plugins|themes}` to `site_transient_update_{plugins|themes}`
 * update `Remote_Management` to work with filter change